server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2 